+++ snorkel/preprocess/core.py
@@ -6,7 +6,7 @@ A preprocessor is a data point to data point mapping in a labeling
 pipeline. This allows LFs to share common preprocessing steps that
 make it easier to express labeling logic. A simple example for text
 processing is concatenating the title and body of an article. For a
-more complex example, see ``snorkel.labeling.preprocess.nlp.SpacyPreprocessor``.
+more complex example, see ``snorkel.preprocess.nlp.SpacyPreprocessor``.
 """
 
 # Used for type checking only
@@ -43,7 +43,7 @@ class preprocessor(lambda_mapper):
     ... def combine_text_preprocessor(x):
     ...     x.article = f"{x.title} {x.body}"
     ...     return x
-    >>> from snorkel.labeling.preprocess.nlp import SpacyPreprocessor
+    >>> from snorkel.preprocess.nlp import SpacyPreprocessor
     >>> spacy_preprocessor = SpacyPreprocessor("article", "article_parsed")
 
     We can now add our preprocessors to an LF.
