+++ test/labeling/apply/test_lf_applier.py
@@ -10,8 +10,8 @@ from dask import dataframe as dd
 from snorkel.labeling.apply import LFApplier, PandasLFApplier
 from snorkel.labeling.apply.dask import DaskLFApplier, PandasParallelLFApplier
 from snorkel.labeling.lf import labeling_function
-from snorkel.labeling.preprocess import preprocessor
-from snorkel.labeling.preprocess.nlp import SpacyPreprocessor
+from snorkel.preprocess import preprocessor
+from snorkel.preprocess.nlp import SpacyPreprocessor
 from snorkel.types import DataPoint
